+ДЗ 5.16. Модуляция. Разбил на модули
